Consolidate scan_path_background_task and scan_path_preprocessing_task into tasks
