+++ pupil_src/shared_modules/scan_path/tasks/__init__.py
@@ -1,0 +1,12 @@
+"""
+(*)~---------------------------------------------------------------------------
+Pupil - eye tracking platform
+Copyright (C) 2012-2020 Pupil Labs
+
+Distributed under the terms of the GNU
+Lesser General Public License (LGPL v3.0).
+See COPYING and COPYING.LESSER for license details.
+---------------------------------------------------------------------------~(*)
+"""
+from .background import ScanPathBackgroundTask
+from .preprocessing import ScanPathPreprocessingTask
