+++ pupil_src/shared_modules/scan_path/tasks/base.py
